Extract common float48, float80 methods
