+++ src/data-buffer.js
@@ -2,8 +2,7 @@
 let debug = () => {}; /* istanbul ignore next */ if (process.env.UTTORI_DATA_DEBUG) { try { debug = require('debug')('DataBuffer'); } catch {} }
 
 const UnderflowError = require('./underflow-error');
-
-// TODO: Extract common helpers to external file between Steam and Buffer: float48. float80, decodeString
+const { float48, float80 } = require('./data-helpers');
 
 /**
  * Helper class for manipulating binary data.
@@ -34,7 +33,7 @@ class DataBuffer {
       debug(error);
       throw new TypeError(error);
     }
-    /** @type {Array|Buffer|Uint8Array} The number of bytes avaliable to read. */
+    /** @type {Array|Buffer|Uint8Array} The bytes avaliable to read. */
     this.data = null;
     if (typeof Buffer !== 'undefined' && Buffer.isBuffer(input)) {
       debug('constructor: from Buffer');
@@ -80,6 +79,9 @@ class DataBuffer {
 
     /** @type {number} Reading / Writing offset */
     this.offset = 0;
+
+    /** @type {number[]} Buffer for creating new files. */
+    this.buffer = [...this.data];
   }
 
   /**
@@ -528,7 +530,7 @@ class DataBuffer {
    */
   readFloat48(littleEndian = false) {
     const uint8 = this.read(6, littleEndian || this.nativeEndian);
-    return this.float48(uint8);
+    return float48(uint8);
   }
 
   /**
@@ -541,7 +543,7 @@ class DataBuffer {
    */
   peekFloat48(offset, littleEndian = false) {
     const uint8 = this.peek(6, offset, littleEndian || this.nativeEndian);
-    return this.float48(uint8);
+    return float48(uint8);
   }
 
   /**
@@ -572,24 +574,24 @@ class DataBuffer {
   /**
    * Read from the current offset and return the IEEE 80 bit extended float value.
    *
-   * @param {boolean} [littleEndian=false] Read in Little Endian format.
+   * @param {boolean} [littleEndian=this.nativeEndian] Read in Little Endian format, defaults to system value.
    * @returns {*} The Float80 value at the current offset.
    */
   readFloat80(littleEndian = this.nativeEndian) {
     const uint8 = this.read(10, littleEndian);
-    return this.float80(uint8);
+    return float80(uint8);
   }
 
   /**
    * Read from the specified offset without advancing the offsets and return the IEEE 80 bit extended float value.
    *
    * @param {number} [offset=0] The offset to read from.
-   * @param {boolean} [littleEndian=false] Read in Little Endian format.
+   * @param {boolean} [littleEndian=this.nativeEndian] Read in Little Endian format, defaults to system value.
    * @returns {*} The Float80 value at the current offset.
    */
   peekFloat80(offset = 0, littleEndian = this.nativeEndian) {
     const uint8 = this.peek(10, offset, littleEndian);
-    return this.float80(uint8);
+    return float80(uint8);
   }
 
   /**
@@ -599,14 +601,11 @@ class DataBuffer {
    * @returns {DataBuffer} The requested number of bytes as a DataBuffer.
    */
   readBuffer(length) {
-    const result = DataBuffer.allocate(length);
-    const to = result.data;
-
+    const to = new Uint8Array(length);
     for (let i = 0; i < length; i++) {
       to[i] = this.readUInt8();
     }
-
-    return result;
+    return new DataBuffer(to);
   }
 
   /**
@@ -617,14 +616,11 @@ class DataBuffer {
    * @returns {DataBuffer} The requested number of bytes as a DataBuffer.
    */
   peekBuffer(offset, length) {
-    const result = DataBuffer.allocate(length);
-    const to = result.data;
-
+    const to = new Uint8Array(length);
     for (let i = 0; i < length; i++) {
       to[i] = this.peekUInt8(offset + i);
     }
-
-    return result;
+    return new DataBuffer(to);
   }
 
   /**
@@ -648,105 +644,6 @@ class DataBuffer {
    */
   peekString(offset, length, encoding = 'ascii') {
     return this.decodeString(offset, length, encoding, false);
-  }
-
-  /**
-   * Convert the current buffer into a Turbo Pascal 48 bit float value.
-   * May be faulty with large numbers due to float percision.
-   *
-   * While most languages use a 32-bit or 64-bit floating point decimal variable, usually called single or double,
-   * Turbo Pascal featured an uncommon 48-bit float called a real which served the same function as a float.
-   *
-   * The Real48 type exists for backward compatibility with Turbo Pascal. It defines a 6-byte floating-point type.
-   * The Real48 type has an 8-bit exponent and a 39-bit normalized mantissa. It cannot store denormalized values, infinity, or not-a-number. If the exponent is zero, the number is zero.
-   *
-   * Structure (Bytes, Big Endian)
-   * 5: SMMMMMMM 4: MMMMMMMM 3: MMMMMMMM 2: MMMMMMMM 1: MMMMMMMM 0: EEEEEEEE
-   *
-   * Structure (Bytes, Little Endian)
-   * 0: EEEEEEEE 1: MMMMMMMM 2: MMMMMMMM 3: MMMMMMMM 4: MMMMMMMM 5: SMMMMMMM
-   *
-   * E[8]: Exponent
-   * M[39]: Mantissa
-   * S[1]: Sign
-   *
-   * Value: (-1)^s * 2^(e - 129) * (1.f)
-   *
-   * @param {Uint8Array} uint8 The data to process to a float48 value.
-   * @returns {number} The read value as a number.
-   * @see {@link http://www.shikadi.net/moddingwiki/Turbo_Pascal_Real|Turbo Pascal Real}
-   */
-  float48(uint8) {
-    debug('float48: uint8', uint8);
-    let mantissa = 0;
-
-    // Bias is 129, which is 0x81
-    let exponent = uint8[0];
-    debug('float48: exponent', exponent);
-    if (exponent === 0) {
-      return 0;
-    }
-    exponent = uint8[0] - 0x81;
-    debug('float48: exponent', exponent);
-
-    for (let i = 1; i <= 4; i++) {
-      mantissa += uint8[i];
-      mantissa /= 256;
-    }
-    mantissa += (uint8[5] & 0x7F);
-    mantissa /= 128;
-    mantissa += 1;
-
-    debug('float48: mantissa', mantissa);
-    // Sign bit check
-    if (uint8[5] & 0x80) {
-      mantissa = -mantissa;
-    }
-    debug('float48: mantissa', mantissa);
-
-    const output = mantissa * (2 ** exponent);
-    debug('float48: output', output);
-    return Number.parseFloat(output.toFixed(4));
-  }
-
-  /**
-   * Convert the current buffer into an IEEE 80 bit extended float value.
-   *
-   * @param {Uint8Array} uint8 The raw data to convert to a float80.
-   * @returns {number} The read value as a number.
-   * @see {@link https://en.wikipedia.org/wiki/Extended_precision|Extended_Precision}
-   */
-  float80(uint8) {
-    const uint32 = new Uint32Array(uint8.buffer, uint8.byteOffset, uint8.byteLength / 4);
-    const [high, low] = [...uint32];
-    const a0 = uint8[9];
-    const a1 = uint8[8];
-
-    // 1 bit sign, -1 or +1
-    const sign = 1 - ((a0 >>> 7) * 2);
-    // 15 bit exponent
-    // let exponent = (((a0 << 1) & 0xFF) << 7) | a1;
-    let exponent = ((a0 & 0x7F) << 8) | a1;
-
-    if ((exponent === 0) && (low === 0) && (high === 0)) {
-      return 0;
-    }
-
-    // 0x7FFF is a reserved value
-    if (exponent === 0x7FFF) {
-      if ((low === 0) && (high === 0)) {
-        return sign * Number.POSITIVE_INFINITY;
-      }
-
-      return Number.NaN;
-    }
-
-    // Bias is 16383, which is 0x3FFF
-    exponent -= 0x3FFF;
-    let out = low * 2 ** (exponent - 31);
-    out += high * 2 ** (exponent - 63);
-
-    return sign * out;
   }
 
   /**
