+++ src/data-stream.js
@@ -3,6 +3,7 @@ let debug = () => {}; /* istanbul ignore next */ if (process.env.UTTORI_DATA_DEB
 const DataBuffer = require('./data-buffer');
 const DataBufferList = require('./data-buffer-list');
 const UnderflowError = require('./underflow-error');
+const { float48, float80 } = require('./data-helpers');
 
 /**
  * Helpter class to ease working with binary files.
@@ -596,7 +597,7 @@ class DataStream {
    */
   readFloat48(littleEndian = false) {
     this.read(6, littleEndian);
-    return this.float48();
+    return float48(this.uint8);
   }
 
   /**
@@ -609,7 +610,7 @@ class DataStream {
    */
   peekFloat48(offset, littleEndian = false) {
     this.peek(6, offset, littleEndian);
-    return this.float48();
+    return float48(this.uint8);
   }
 
   /**
@@ -643,7 +644,7 @@ class DataStream {
    */
   readFloat80(littleEndian = false) {
     this.read(10, littleEndian);
-    return this.float80();
+    return float80(this.uint8);
   }
 
   /**
@@ -655,7 +656,7 @@ class DataStream {
    */
   peekFloat80(offset = 0, littleEndian = false) {
     this.peek(10, offset, littleEndian);
-    return this.float80();
+    return float80(this.uint8);
   }
 
   /**
@@ -665,14 +666,11 @@ class DataStream {
    * @returns {DataBuffer} The requested number of bytes as a DataBuffer.
    */
   readBuffer(length) {
-    const result = DataBuffer.allocate(length);
-    const to = result.data;
-
+    const to = new Uint8Array(length);
     for (let i = 0; i < length; i++) {
       to[i] = this.readUInt8();
     }
-
-    return result;
+    return new DataBuffer(to);
   }
 
   /**
@@ -683,14 +681,11 @@ class DataStream {
    * @returns {DataBuffer} The requested number of bytes as a DataBuffer.
    */
   peekBuffer(offset, length) {
-    const result = DataBuffer.allocate(length);
-    const to = result.data;
-
+    const to = new Uint8Array(length);
     for (let i = 0; i < length; i++) {
       to[i] = this.peekUInt8(offset + i);
     }
-
-    return result;
+    return new DataBuffer(to);
   }
 
   /**
@@ -739,96 +734,6 @@ class DataStream {
    */
   peekString(offset, length, encoding = 'ascii') {
     return this.decodeString(offset, length, encoding, false);
-  }
-
-  /**
-   * Convert the current buffer into a Turbo Pascal 48 bit float value.
-   * May be faulty with large numbers due to float percision.
-   *
-   * While most languages use a 32-bit or 64-bit floating point decimal variable, usually called single or double,
-   * Turbo Pascal featured an uncommon 48-bit float called a real which served the same function as a float.
-   *
-   * The Real48 type exists for backward compatibility with Turbo Pascal. It defines a 6-byte floating-point type.
-   * The Real48 type has an 8-bit exponent and a 39-bit normalized mantissa. It cannot store denormalized values, infinity, or not-a-number. If the exponent is zero, the number is zero.
-   *
-   * Structure (Bytes, Big Endian)
-   * 5: SMMMMMMM 4: MMMMMMMM 3: MMMMMMMM 2: MMMMMMMM 1: MMMMMMMM 0: EEEEEEEE
-   *
-   * Structure (Bytes, Little Endian)
-   * 0: EEEEEEEE 1: MMMMMMMM 2: MMMMMMMM 3: MMMMMMMM 4: MMMMMMMM 5: SMMMMMMM
-   *
-   * E[8]: Exponent
-   * M[39]: Mantissa
-   * S[1]: Sign
-   *
-   * Value: (-1)^s * 2^(e - 129) * (1.f)
-   *
-   * @returns {number} The read value as a number.
-   * @see {@link http://www.shikadi.net/moddingwiki/Turbo_Pascal_Real|Turbo Pascal Real}
-   */
-  float48() {
-    let mantissa = 0;
-
-    // Bias is 129, which is 0x81
-    let exponent = this.uint8[0];
-    if (exponent === 0) {
-      return 0;
-    }
-    exponent = this.uint8[0] - 0x81;
-
-    for (let i = 1; i <= 4; i++) {
-      mantissa += this.uint8[i];
-      mantissa /= 256;
-    }
-    mantissa += (this.uint8[5] & 0x7F);
-    mantissa /= 128;
-    mantissa += 1;
-
-    // Sign bit check
-    if (this.uint8[5] & 0x80) {
-      mantissa = -mantissa;
-    }
-
-    const output = mantissa * (2 ** exponent);
-    return Number.parseFloat(output.toFixed(4));
-  }
-
-  /**
-   * Convert the current buffer into an IEEE 80 bit extended float value.
-   *
-   * @returns {number} The read value as a number.
-   * @see {@link https://en.wikipedia.org/wiki/Extended_precision|Extended_Precision}
-   */
-  float80() {
-    const [high, low] = [...this.uint32];
-    const a0 = this.uint8[9];
-    const a1 = this.uint8[8];
-
-    // 1 bit sign, -1 or +1
-    const sign = 1 - ((a0 >>> 7) * 2);
-    // 15 bit exponent
-    // let exponent = (((a0 << 1) & 0xFF) << 7) | a1;
-    let exponent = ((a0 & 0x7F) << 8) | a1;
-
-    if ((exponent === 0) && (low === 0) && (high === 0)) {
-      return 0;
-    }
-
-    // 0x7FFF is a reserved value
-    if (exponent === 0x7FFF) {
-      if ((low === 0) && (high === 0)) {
-        return sign * Number.POSITIVE_INFINITY;
-      }
-
-      return Number.NaN;
-    }
-
-    // Bias is 16383, which is 0x3FFF
-    exponent -= 0x3FFF;
-    let out = low * 2 ** (exponent - 31);
-    out += high * 2 ** (exponent - 63);
-
-    return sign * out;
   }
 
   /**
