+++ test/check-tree-shaking.test.js
@@ -35,6 +35,7 @@ test('Tree Shaking: { DataBuffer, DataBufferList, DataStream }', async (t) => {
   // Sum should be 1 + number of expected modules
   t.deepEqual(Object.keys(output.output[0].modules).map((f) => path.basename(f)), [
     'underflow-error.js',
+    'data-helpers.js',
     'data-buffer.js',
     'data-buffer-list.js',
     'data-stream.js',
@@ -56,6 +57,7 @@ test('Tree Shaking: { DataBitstream }', async (t) => {
   // Sum should be 1 + number of expected modules
   t.deepEqual(Object.keys(output.output[0].modules).map((f) => path.basename(f)), [
     'underflow-error.js',
+    'data-helpers.js',
     'data-buffer.js',
     'data-buffer-list.js',
     'data-stream.js',
@@ -78,6 +80,7 @@ test('Tree Shaking: { CRC32, LZW }', async (t) => {
   // Sum should be 1 + number of expected modules
   t.deepEqual(Object.keys(output.output[0].modules).map((f) => path.basename(f)), [
     'underflow-error.js',
+    'data-helpers.js',
     'data-buffer.js',
     'data-hash-crc32.js',
     'data-compression-lzw.js',
